Make converting from JSON to DSON works
Also change the data types to `int32` and `int` accordingly:

- `int32` for "raw" data
- `int` for inferred data

+++ dson/dfield/infer.go
@@ -16,9 +16,9 @@ import (
 )
 
 func InferUsingMeta2Entry(rawData []byte, meta2Entry dmeta2.Entry) Inferences {
-	rawDataOffset := meta2Entry.Offset + meta2Entry.Inferences.FieldNameLength
+	rawDataOffset := int(meta2Entry.Offset) + meta2Entry.Inferences.FieldNameLength
 	rawDataLength := meta2Entry.Inferences.RawDataLength
-	alignedBytesCount := int32(ds.NearestDivisibleByM(int(rawDataOffset), 4)) - rawDataOffset
+	alignedBytesCount := ds.NearestDivisibleByM(rawDataOffset, 4) - rawDataOffset
 	rawDataStripped := rawData
 	if rawDataLength > alignedBytesCount {
 		rawDataStripped = rawData[alignedBytesCount:]
@@ -36,7 +36,7 @@ func InferUsingMeta2Entry(rawData []byte, meta2Entry dmeta2.Entry) Inferences {
 	}
 }
 
-func InferHierarchyPath(index int32, fields []DataField) []string {
+func InferHierarchyPath(index int, fields []DataField) []string {
 	// TODO: create a cache function if there is need for optimization
 	fieldName := fields[index].Name
 	parentIndex := fields[index].Inferences.ParentIndex
@@ -50,7 +50,7 @@ func InferHierarchyPaths(fields []DataField) []DataField {
 	fieldsCopy := lo.Map(
 		fields,
 		func(t DataField, i int) DataField {
-			t.Inferences.HierarchyPath = InferHierarchyPath(int32(i), fields)
+			t.Inferences.HierarchyPath = InferHierarchyPath(i, fields)
 			return t
 		},
 	)
